Callback controller added

+++ src/main/java/cc/codedhyan/codeitup/problem/model/TestCases.java
@@ -34,7 +34,7 @@ public class TestCases {
 
     private String time;
 
-    private Double memory;
+    private Integer memory;
 
     @Enumerated(EnumType.STRING)
     private TestCasesResult testCasesResult;
